Move sharedcomponent to /internal
