+++ internal/sharedcomponent/sharedcomponent_test.go
